Added ClientsList and MemberList. Updated generateID() in TheaterPatron. Made Theater constructor work as a singleton. Worked on Client.hasDate().
NOTES: I spent a lot of time playing with a TheaterList parent class for
the two collections, but it wasn't working out. Obviously, the parent
can't be a singleton, because each child needs a copy. I couldn't get
the generic aspects of the parent class to work right. If anyone else
wants to try it, feel free. I'm not sure it's necessary.

After winding down for the night, I realized that Client.hasDate() works
fine for a single date, but isn't going to cut it for a new Show that is
more than one day. This method needs revisited. I think the principle is
sound, but the implementation is lacking.